:wrench: replace filterServices logic with ES queries and addMessages
